refactor(strapi addons): Use a folder with low privilege for downloads
