+++ backend/src/plugins/addons/server/services/seed/backup/export.ts
@@ -3,23 +3,7 @@ import {
    ensureDirectoryExists,
    isStderrAnError,
 } from '../../../helpers/server-helpers';
-
-export const BACKUP_FILE_NAME = 'export';
-
-const BACKUP_FOLDER_NAME = 'backup';
-
-const BACKUP_FOLDER_PATH = `public/${BACKUP_FOLDER_NAME}`;
-
-const EXPORT_FILE_PATH = `${BACKUP_FOLDER_PATH}/${BACKUP_FILE_NAME}`;
-
-interface GetBackupPathInput {
-   isEncrypted: boolean;
-}
-
-export function getBackupPath({ isEncrypted }: GetBackupPathInput) {
-   const exportFilePath = `${BACKUP_FOLDER_NAME}/${BACKUP_FILE_NAME}.tar.gz`;
-   return isEncrypted ? `${exportFilePath}.enc` : exportFilePath;
-}
+import { BACKUP_FOLDER_PATH, EXPORT_FILE_PATH } from './utils';
 
 export async function exportBackup(): Promise<string> {
    await ensureDirectoryExists(BACKUP_FOLDER_PATH);
